+++ src/main.ts
@@ -61,7 +61,7 @@ export async function render(progress: SetProgressFn, assets: string, file: stri
 				throw new Error(`PDF render error`);
 			const files = fs.readdirSync(tmpfolder).sort().map(f => path.join(tmpfolder, f));
 			const fout = path.join(fdir, fname + ".pdf");
-			await mergePDFs(files, fout, doc.title, doc.author);
+			await mergePDFs(files, fout, doc);
 			fs.rmSync(tmpfolder, { recursive: true, force: true });
 			return fout;
 		}
@@ -128,11 +128,13 @@ function runDocxMacro(progress: SetProgressFn, assets: string, cwd: string, fin:
 }
 
 
-async function mergePDFs(files: string[], fout: string, title: string | undefined, author: string | undefined)
+async function mergePDFs(files: string[], fout: string, doc?: Doc)
 {
 	const mergedPdf = await PDFDocument.create();
-	if (title) mergedPdf.setTitle(title, { showInWindowTitleBar: true });
-	if (author) mergedPdf.setAuthor(author);
+	if (doc?.title) mergedPdf.setTitle(doc.title, { showInWindowTitleBar: true });
+	if (doc?.author) mergedPdf.setAuthor(doc.author);
+	if (doc?.ctime) mergedPdf.setCreationDate(doc.ctime);
+	if (doc?.mtime) mergedPdf.setModificationDate(doc.mtime);
 	for (const file of files)
 	{
 		const buffer = fs.readFileSync(file);